Remove fideslang module in favor of fideslang package
