+++ requirements.txt
@@ -3,6 +3,7 @@ click>=8.1
 colorama>=0.4.3
 cryptography>=36
 deepdiff==5.8.0
+fideslang==0.9.0
 fideslog==1.1.5
 loguru>=0.5,<0.6
 openpyxl==3.0.9
