+++ src/fidesapi/routes/visualize.py
@@ -6,11 +6,11 @@ from typing import Union
 
 from fastapi import APIRouter, HTTPException
 from fastapi.responses import HTMLResponse
+from fideslang import model_map
 
 from fidesapi.routes.crud import list_resource
 from fidesapi.sql_models import sql_model_map
 from fidesctl.core import visualize
-from fideslang import model_map
 
 # pylint: disable=redefined-outer-name,cell-var-from-loop
