+++ src/fidesctl/cli/options.py
@@ -5,7 +5,6 @@ Contains all of the options/arguments used by the CLI commands.
 from typing import Callable
 
 import click
-
 from fideslang import model_list
