+++ src/fidesctl/core/api_helpers.py
@@ -4,12 +4,12 @@ Reusable utilities meant to make repetitive api-related tasks easier.
 
 from typing import Dict, List, Optional
 
-from requests import Response
-
-from fidesctl.core import api
 from fideslang import FidesModel
 from fideslang.parse import parse_dict
 from fideslang.validation import FidesKey
+from requests import Response
+
+from fidesctl.core import api
 
 
 def get_server_resources(
