+++ src/fidesctl/core/apply.py
@@ -4,12 +4,12 @@ from pprint import pprint
 from typing import Dict, List, Tuple
 
 from deepdiff import DeepDiff
+from fideslang import FidesModel, Taxonomy
 
 from fidesctl.cli.utils import handle_cli_response
 from fidesctl.core import api
 from fidesctl.core.api_helpers import get_server_resources
 from fidesctl.core.utils import echo_green, echo_red
-from fideslang import FidesModel, Taxonomy
 
 
 def sort_create_update(
