+++ src/fidesctl/core/parse.py
@@ -1,8 +1,9 @@
 """This module is responsible for parsing and verifying file, either with or without a server being available."""
-from fidesctl.core.utils import echo_green
 from fideslang import Taxonomy
 from fideslang.manifests import ingest_manifests
 from fideslang.parse import load_manifests_into_taxonomy
+
+from fidesctl.core.utils import echo_green
 
 
 def parse(manifests_dir: str) -> Taxonomy:
