+++ tests/core/test_api_helpers.py
@@ -3,11 +3,11 @@ import uuid
 from typing import TYPE_CHECKING, Dict, Generator, List, Optional
 
 import pytest
+from fideslang import FidesModel, model_list
 
 from fidesctl.core import api as _api
 from fidesctl.core import api_helpers as _api_helpers
 from fidesctl.core.config import FidesctlConfig
-from fideslang import FidesModel, model_list
 
 RESOURCE_CREATION_COUNT = 5
 # These resources have tricky validation so the fides_key replacement doesn't work
